requirements.txt
Flask==3.0.3
supabase==2.4.6
python-dotenv==1.0.1
Werkzeug==3.0.3
PyJWT==2.8.0
requests==2.31.0
